Secures most pages, adds user json to output html
Most urls routed through the Application controller are now secured, and
redirect to the index if an unauthenticated/unauthorized user attempts to
access them.

User information is rendered to the html output as a json object, so we can
easily see at the JS level if we have a current user.